Implement reserved IPv4 addresses for IMSIs
Allows maintaining, via the VTY, a list of IMSI -> IPv4
address mappings to ensure the GGSN always assigns
the same IP address to a UE identified by the IMSI.

This is NOT an implenentation of static IP address support,
it simply reserves IPs from the dynamic IP pool. No checks
are done to enure that the address indeed forms part of the
defined pool, this is left up to the operator to configure
correctly.

It may be desired to always have the same IP address assigned
after a PDP context is deleted and recreated by the device,
by toggling mobile data, or airplane mode.
This could be useful, for example, in order to avoid having to
adjust firewall or packet capture filters. One might also have
a small subset of IMSIs on a network that should have a special
set of firewall rules, for access to internal infrastructure or
traffic shaping exemptions, as some examples.

This is intented as a developer or small operator feature, but
it is not the intention here that one would go about assigning
'static' IP addresses to an entire HLR of subscribers via the
GGSN vty.