Move in46_addr to libosmocore
I'm in the process of extending the public API of OsmoGGSN's libgtp, so
a GSN can be created with either IPv4 or IPv6. The in_addr code from
osmo-ggsn.git provides a good abstraction for that. I've considered
making it public in libgtp, but given that it is generic code that could
be used outside of libgtp in other Osmocom projects, put it in
libosmocore instead.

Depends: libosmocore I31078f130ec42aeaa5ead1dde82fdd1eb44d992b
Change-Id: I8f4d9d78689909a149683583cfe36bcdc3f7cbc7