Rename tun_encaps -> tun_inject_pkt
The terminology of "encapsulating" used in tun_encaps is confusing,
since in this case we are not really encapsulating anything, but
actually delivering ("injecting") the packet to the system network
stack.
Using the "encapsulating" term is also confusing because readers may
think it may be doing the GTP-U encapsulation, which is not.

Change-Id: Ibb17c792b81668571e69d79918f3adf0e9e0f6c8